requirements.txt
# PharmOps AI Assistant — direct dependencies only
# pip install -r requirements.txt will automatically install all sub-dependencies

# LangChain — agent framework and RAG pipeline
langchain
langchain-community
langchain-google-genai
langchain-text-splitters

# Vector store — local FAISS index for SOP embeddings
faiss-cpu

# LLM and embeddings — Google Gemini free tier
# Confirmed working embedding model: models/gemini-embedding-001
# Confirmed working LLM model: gemini-1.5-flash
google-genai

# UI — Streamlit chat interface
streamlit

# Database — Snowflake live data queries
snowflake-connector-python

# MCP — Model Context Protocol server for tool exposure
mcp

# Utilities
python-dotenv